Issue #9 Reoccuring updates
renamed file from initial_loading.py to update-neo4j.py
created config file for storing variables for scripts
updates database from last run of script, if it has not been ran it retrieves the last 50 years of articles from the current date

- Code needs to be added for article removal for articles that fall past the 50 years
