+++ config.ini
@@ -1,0 +1,13 @@
+[credentials]
+neo4j_username = 
+neo4j_password = 
+neo4j_uri = 
+omim_api_key = 
+
+[database]
+database_last_run = 
+
+[script]
+from_disease = 0
+to_disease = 2000
+
